backend/requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
python-dotenv==1.0.0
google-generativeai==0.3.0
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1